Split up UI elements into separate files, create trait for update
